Namespace refactoring for SonarQube
