+++ CorePageFactory.sln
@@ -5,15 +5,19 @@ VisualStudioVersion = 15.0.28010.2036
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "WebDriverFactory", "WebDriverFactory", "{178E49E3-799F-4FF1-A3B2-154DB4551E00}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AlexanderOnTest.WebDriverFactory", "WebDriverFactory\src\AlexanderOnTest.WebDriverFactory\AlexanderOnTest.WebDriverFactory.csproj", "{7370C31D-F157-4F94-9EEE-4AB335802F41}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AlexanderOnTest.WebDriverFactory.WindowsTests", "WebDriverFactory\test\AlexanderOnTest.WebDriverFactory.WindowsTests\AlexanderOnTest.WebDriverFactory.WindowsTests.csproj", "{E37C9341-7092-499E-B58F-4B401B17EA26}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AlexanderOnTest.WebDriverFactory.LinuxTests", "WebDriverFactory\test\AlexanderOnTest.WebDriverFactory.LinuxTests\AlexanderOnTest.WebDriverFactory.LinuxTests.csproj", "{52DC6236-753F-4135-B579-555BA5E98813}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AlexanderOnTest.WebDriverFactory.MacOsTests", "WebDriverFactory\test\AlexanderOnTest.WebDriverFactory.MacOsTests\AlexanderOnTest.WebDriverFactory.MacOsTests.csproj", "{FA1007A4-BDF1-4E34-9F41-F41CF5AACC12}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{30D9B787-E07B-4FAA-8F50-87A5E4A18DE3}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AoT.WebDriverFactory", "src\WebDriverFactory\AoT.WebDriverFactory\AoT.WebDriverFactory.csproj", "{4A6FF976-F25E-420D-A238-6BAB8CC860A5}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AoT.WebDriverFactory.WindowsTests", "test\WebDriverFactory\AoT.WebDriverFactory.WindowsTests\AoT.WebDriverFactory.WindowsTests.csproj", "{4688105F-54C7-42FA-8218-D36BEB35DFDD}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "WebDriverFactory", "WebDriverFactory", "{B41D6301-B590-4AF9-A93F-3CC9DAF65B90}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{567247CB-5AAB-4638-A4B3-DEDC66B6EC7F}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AoT.WebDriverFactory.MacOsTests", "test\WebDriverFactory\AoT.WebDriverFactory.MacOsTests\AoT.WebDriverFactory.MacOsTests.csproj", "{5723B399-BD1B-45BE-9562-9A9E5D63FE23}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AoT.WebDriverFactory.LinuxTests", "test\WebDriverFactory\AoT.WebDriverFactory.LinuxTests\AoT.WebDriverFactory.LinuxTests.csproj", "{0BB2A572-E7F1-4533-9E24-FE4C73B03242}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,31 +25,33 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7370C31D-F157-4F94-9EEE-4AB335802F41}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7370C31D-F157-4F94-9EEE-4AB335802F41}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7370C31D-F157-4F94-9EEE-4AB335802F41}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7370C31D-F157-4F94-9EEE-4AB335802F41}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E37C9341-7092-499E-B58F-4B401B17EA26}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E37C9341-7092-499E-B58F-4B401B17EA26}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E37C9341-7092-499E-B58F-4B401B17EA26}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E37C9341-7092-499E-B58F-4B401B17EA26}.Release|Any CPU.Build.0 = Release|Any CPU
-		{52DC6236-753F-4135-B579-555BA5E98813}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{52DC6236-753F-4135-B579-555BA5E98813}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{52DC6236-753F-4135-B579-555BA5E98813}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{52DC6236-753F-4135-B579-555BA5E98813}.Release|Any CPU.Build.0 = Release|Any CPU
-		{FA1007A4-BDF1-4E34-9F41-F41CF5AACC12}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FA1007A4-BDF1-4E34-9F41-F41CF5AACC12}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FA1007A4-BDF1-4E34-9F41-F41CF5AACC12}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FA1007A4-BDF1-4E34-9F41-F41CF5AACC12}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4A6FF976-F25E-420D-A238-6BAB8CC860A5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4A6FF976-F25E-420D-A238-6BAB8CC860A5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4A6FF976-F25E-420D-A238-6BAB8CC860A5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4A6FF976-F25E-420D-A238-6BAB8CC860A5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4688105F-54C7-42FA-8218-D36BEB35DFDD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4688105F-54C7-42FA-8218-D36BEB35DFDD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4688105F-54C7-42FA-8218-D36BEB35DFDD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4688105F-54C7-42FA-8218-D36BEB35DFDD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5723B399-BD1B-45BE-9562-9A9E5D63FE23}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5723B399-BD1B-45BE-9562-9A9E5D63FE23}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5723B399-BD1B-45BE-9562-9A9E5D63FE23}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5723B399-BD1B-45BE-9562-9A9E5D63FE23}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0BB2A572-E7F1-4533-9E24-FE4C73B03242}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0BB2A572-E7F1-4533-9E24-FE4C73B03242}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0BB2A572-E7F1-4533-9E24-FE4C73B03242}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0BB2A572-E7F1-4533-9E24-FE4C73B03242}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{7370C31D-F157-4F94-9EEE-4AB335802F41} = {178E49E3-799F-4FF1-A3B2-154DB4551E00}
-		{E37C9341-7092-499E-B58F-4B401B17EA26} = {30D9B787-E07B-4FAA-8F50-87A5E4A18DE3}
-		{52DC6236-753F-4135-B579-555BA5E98813} = {30D9B787-E07B-4FAA-8F50-87A5E4A18DE3}
-		{FA1007A4-BDF1-4E34-9F41-F41CF5AACC12} = {30D9B787-E07B-4FAA-8F50-87A5E4A18DE3}
+		{178E49E3-799F-4FF1-A3B2-154DB4551E00} = {567247CB-5AAB-4638-A4B3-DEDC66B6EC7F}
+		{4A6FF976-F25E-420D-A238-6BAB8CC860A5} = {178E49E3-799F-4FF1-A3B2-154DB4551E00}
+		{4688105F-54C7-42FA-8218-D36BEB35DFDD} = {B41D6301-B590-4AF9-A93F-3CC9DAF65B90}
+		{B41D6301-B590-4AF9-A93F-3CC9DAF65B90} = {30D9B787-E07B-4FAA-8F50-87A5E4A18DE3}
+		{5723B399-BD1B-45BE-9562-9A9E5D63FE23} = {B41D6301-B590-4AF9-A93F-3CC9DAF65B90}
+		{0BB2A572-E7F1-4533-9E24-FE4C73B03242} = {B41D6301-B590-4AF9-A93F-3CC9DAF65B90}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {064D603D-68B4-4F0E-A629-21262CC85724}
