+++ src/WebDriverFactory/AoT.WebDriverFactory/IDriverOptionsFactory.cs
@@ -1,14 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using OpenQA.Selenium;
+﻿using OpenQA.Selenium;
 using OpenQA.Selenium.Chrome;
 using OpenQA.Selenium.Edge;
 using OpenQA.Selenium.Firefox;
 using OpenQA.Selenium.IE;
 using OpenQA.Selenium.Safari;
 
-namespace AlexanderOnTest.WebDriverFactory
+namespace AoT.WebDriverFactory
 {
     public interface IDriverOptionsFactory
     {
