+++ test/WebDriverFactory/AoT.WebDriverFactory.LinuxTests/StaticWebDriverFactoryTests.cs
@@ -3,9 +3,8 @@ using System.Drawing;
 using FluentAssertions;
 using NUnit.Framework;
 using OpenQA.Selenium;
-using OpenQA.Selenium.Firefox;
 
-namespace AlexanderOnTest.WebDriverFactory.LinuxTests
+namespace AoT.WebDriverFactory.LinuxTests
 {
     [TestFixture]
     public class StaticWebDriverFactoryTests
