+++ test/WebDriverFactory/AoT.WebDriverFactory.WindowsTests/AoT.WebDriverFactory.WindowsTests.csproj
@@ -17,11 +17,13 @@
     <PackageReference Include="Selenium.Support" Version="3.14.0" />
     <PackageReference Include="Selenium.WebDriver" Version="3.14.0" />
     <PackageReference Include="Selenium.WebDriver.ChromeDriver" Version="2.42.0.1" />
-    <PackageReference Include="Selenium.WebDriver.GeckoDriver" Version="0.22.0" />
+    <PackageReference Include="Selenium.WebDriver.GeckoDriver.Win64" Version="0.22.0" />
+    <PackageReference Include="Selenium.WebDriver.IEDriver" Version="3.14.0" />
+    <PackageReference Include="Selenium.WebDriver.MicrosoftDriver" Version="17.17134.0" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\AlexanderOnTest.WebDriverFactory\AlexanderOnTest.WebDriverFactory.csproj" />
+    <ProjectReference Include="..\..\..\src\WebDriverFactory\AoT.WebDriverFactory\AoT.WebDriverFactory.csproj" />
   </ItemGroup>
 
 </Project>
