+++ test/WebDriverFactory/AoT.WebDriverFactory.WindowsTests/WebDriverFactoryInstanceTests.cs
@@ -3,20 +3,21 @@ using System.Drawing;
 using FluentAssertions;
 using NUnit.Framework;
 using OpenQA.Selenium;
-using OpenQA.Selenium.Firefox;
 
-namespace AlexanderOnTest.WebDriverFactory.WindowsTests
+namespace AoT.WebDriverFactory.WindowsTests
 {
     [TestFixture]
-    public class StaticWebDriverFactoryTests
+    public class WebDriverFactoryInstanceTests
     {
         private IWebDriver Driver { get; set; }
         private readonly PlatformType thisPlatformType = PlatformType.Windows;
+        private IWebDriverFactory WebDriverFactory { get; set; }
 
         [OneTimeSetUp]
-        public void CheckForValidPlatform()
+        public void SetUp()
         {
             Assume.That(() => Platform.CurrentPlatform.IsPlatformType(thisPlatformType));
+            this.WebDriverFactory = new DefaultWebDriverFactory();
         }
 
         [Test]
@@ -26,7 +27,7 @@ namespace AlexanderOnTest.WebDriverFactory.WindowsTests
         [TestCase(Browser.Chrome)]
         public void LocalWebDriverCanBeLaunchedAndLoadExampleDotCom(Browser browser)
         {
-            Driver = StaticWebDriverFactory.GetLocalWebDriver(browser);
+            Driver = this.WebDriverFactory.GetLocalWebDriver(browser);
             Driver.Url = "https://example.com/";
             Driver.Title.Should().Be("Example Domain");
         }
@@ -35,7 +36,7 @@ namespace AlexanderOnTest.WebDriverFactory.WindowsTests
         [TestCase(Browser.Safari)]
         public void RequestingUnsupportedWebDriverThrowsInformativeException(Browser browser)
         {
-            Action act = () => StaticWebDriverFactory.GetLocalWebDriver(browser);
+            Action act = () => this.WebDriverFactory.GetLocalWebDriver(browser);
             act.Should()
                 .Throw<PlatformNotSupportedException>($"because {browser} is not supported on {thisPlatformType}.")
                 .WithMessage("*is only available on*");
@@ -46,7 +47,7 @@ namespace AlexanderOnTest.WebDriverFactory.WindowsTests
         [TestCase(Browser.Chrome)]
         public void HeadlessBrowsersCanBeLaunched(Browser browser)
         {
-            Driver = StaticWebDriverFactory.GetLocalWebDriver(browser, true);
+            Driver = this.WebDriverFactory.GetLocalWebDriver(browser, true);
             Driver.Url = "https://example.com/";
             Driver.Title.Should().Be("Example Domain");
         }
@@ -57,7 +58,7 @@ namespace AlexanderOnTest.WebDriverFactory.WindowsTests
         [TestCase(Browser.Safari)]
         public void RequestingUnsupportedHeadlessBrowserThrowsInformativeException(Browser browser)
         {
-            Action act = () => StaticWebDriverFactory.GetLocalWebDriver(browser, true);
+            Action act = () => this.WebDriverFactory.GetLocalWebDriver(browser, true);
             act.Should()
                 .ThrowExactly<ArgumentException>($"because headless mode is not supported on {browser}.")
                 .WithMessage($"Headless mode is not currently supported for {browser}.");
@@ -66,7 +67,7 @@ namespace AlexanderOnTest.WebDriverFactory.WindowsTests
         [Test]
         public void HdBrowserIsOfRequestedSize()
         {
-            Driver = StaticWebDriverFactory.GetLocalWebDriver(StaticDriverOptionsFactory.GetFirefoxOptions(true), WindowSize.Hd);
+            Driver = this.WebDriverFactory.GetLocalWebDriver(StaticDriverOptionsFactory.GetFirefoxOptions(true), WindowSize.Hd);
 
             Assert.Multiple(() =>
             {
@@ -79,7 +80,7 @@ namespace AlexanderOnTest.WebDriverFactory.WindowsTests
         [Test]
         public void FhdBrowserIsOfRequestedSize()
         {
-            Driver = StaticWebDriverFactory.GetLocalWebDriver(StaticDriverOptionsFactory.GetFirefoxOptions(true), WindowSize.Fhd);
+            Driver = this.WebDriverFactory.GetLocalWebDriver(StaticDriverOptionsFactory.GetFirefoxOptions(true), WindowSize.Fhd);
 
             Assert.Multiple(() =>
             {
@@ -96,7 +97,7 @@ namespace AlexanderOnTest.WebDriverFactory.WindowsTests
         [TestCase(Browser.Chrome)]
         public void RemoteWebDriverCanBeLaunchedAndLoadExampleDotCom(Browser browser)
         {
-            Driver = StaticWebDriverFactory.GetRemoteWebDriver(browser, new Uri("http://localhost:4444/wd/hub"), PlatformType.Any);
+            Driver = this.WebDriverFactory.GetRemoteWebDriver(browser);
             Driver.Url = "https://example.com/";
             Driver.Title.Should().Be("Example Domain");
         }
